requirements.txt
streamlit>=1.44.0
pandas>=2.2.0
requests>=2.32.0
beautifulsoup4>=4.12.0
python-dotenv>=1.0.0
pypdf>=6.0.0
openpyxl>=3.1.0
reportlab>=4.0.0
